Show Modal If Profile Incomplete
- show dialog when dashboard mounts
- fix isProfileComplete boolean logic
