+++ src/pages/Dashboard/DashboardLayout.tsx
@@ -1,8 +1,12 @@
+import { ReactNode, useContext, useEffect, useRef } from "react";
+import styles from "./DashboardLayout.module.scss";
+
+import { UserContext } from "../../contexts/UserContext";
+
 import Icon from "../../components/common/Icon";
 import Navbar from "../../components/common/Navbar/Navbar";
-import { UserContext } from "../../contexts/UserContext";
-import styles from "./DashboardLayout.module.scss";
-import { ReactNode, useContext } from "react";
+import { Button } from "../../components/common/Button/Button";
+import { useLocation, useNavigate } from "react-router-dom";
 
 const dashboardLinks = [
 	{ href: "/dashboard", name: "Links", icon: <Icon variant='link' /> },
@@ -19,13 +23,50 @@ const dashboardLinks = [
 ];
 
 const DashboardLayout = ({ children }: { children: ReactNode }) => {
-	const { isUserPending } = useContext(UserContext);
+	const { isUserPending, isProfileComplete } = useContext(UserContext);
+	const navigate = useNavigate();
+	const location = useLocation();
+
+	const dialogRef = useRef<HTMLDialogElement>(null);
+
+	const handleCloseDialog = () => {
+		dialogRef.current?.close();
+	};
+
+	useEffect(() => {
+		const mainDashboardPage = location.pathname === "/dashboard";
+		if (!isProfileComplete && mainDashboardPage) {
+			dialogRef.current?.showModal();
+		}
+	}, [isProfileComplete, location.pathname]);
 
 	return (
 		<div className={styles.dashboard_layout}>
 			<div>
 				<Navbar navigationLinks={dashboardLinks} className={styles.navbar} />
-				<main>{isUserPending ? <h1>Loading...</h1> : children}</main>
+				<main>
+					{isUserPending ? <h1>Loading...</h1> : children}
+					{!isProfileComplete && (
+						<dialog ref={dialogRef}>
+							<div className={styles.dialog__content}>
+								<div>
+									<h2>Please complete your profile</h2>
+								</div>
+								<div className={styles.dialog__actionBtns}>
+									<Button
+										variant='default'
+										onClick={() => navigate("/dashboard/profile")}
+									>
+										Profile
+									</Button>
+									<Button variant='secondary' onClick={handleCloseDialog}>
+										Skip
+									</Button>
+								</div>
+							</div>
+						</dialog>
+					)}
+				</main>
 			</div>
 		</div>
 	);
